(executor-api) Make executor API talk GRPC
The executor's REST API was very fragile and annoying to work with, lacking even basic type safety.  Migrate to use GRPC instead.  GRPC is a bit of a pain with how verbose it is, but that is probably a lesser evil.  This is a fairly straightforward change, but it's also large so a solid round of testing is needed...

The change set breaks out the GrpcStubPool previously residing in the QueryService, and makes it available to all clients.

ServiceId.name was also renamed to avoid the very dangerous clash with Enum.name().

The boilerplate needed for grpc was also extracted into a common gradle file for inclusion into the appropriate build.gradle-files.

+++ code/common/config/src/main/java/nu/marginalia/storage/model/FileStorageId.java
@@ -4,7 +4,7 @@ public record FileStorageId(long id) {
     public static FileStorageId parse(String str) {
         return new FileStorageId(Long.parseLong(str));
     }
-    public static FileStorageId of(int storageId) {
+    public static FileStorageId of(long storageId) {
         return new FileStorageId(storageId);
     }
